Move gallery items into /pages
