--- app/javascript/components/Faq.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-
-class Faq extends React.Component {
-
-  render() {
-
-  }
-
-}
-
-export default Faq;
