+++ app/javascript/index.jsx
@@ -9,7 +9,7 @@ import '../assets/styles/app.scss';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { Route } from 'react-router-dom';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { ConnectedRouter } from 'react-router-redux';
 
 import { history, store } from './store';
@@ -20,7 +20,9 @@ import App from './containers/App';
 const Index = () => (
   <Provider store={store}>
     <ConnectedRouter history={history}>
-      <Route path="/" component={App} />
+      <Router>
+        <Route path="/" component={App} />
+      </Router>
     </ConnectedRouter>
   </Provider>
 );
