+++ app/javascript/pages/components/Gallery.jsx
