+++ app/javascript/pages/components/PrivateRoute.jsx
