+++ app/javascript/styles/components/Login.scss
@@ -1,0 +1,22 @@
+.Login {
+  align-content: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.login {
+  &__frame {
+    margin-top: 100px;
+    width: 300px;
+  }
+
+  &__form {
+    display: flex;
+    flex-direction: column;
+  }
+
+  &__form > .submit {
+      margin-top: 1rem;
+  }
+}
