+++ app/javascript/styles/components/YearNav.scss
@@ -1,0 +1,77 @@
+.year-navigation {
+  background: $color_bg-medium;
+  box-shadow: 3px 3px 6px rgba(0, 0, 0, .16);
+  height: 80px;
+  margin-bottom: 60px;
+
+  &__mobile {
+    display: none;
+  }
+
+  &__list {
+    align-items: center;
+    display: flex;
+    flex-direction: row;
+    height: 80px;
+    justify-content: space-around;
+  }
+
+  &__item {
+    @include media(medium-small) { width: 45px; }
+    width: 70px;
+  }
+
+  &__year {
+    @include media(medium-small) { font-size: $font_size-medium; }
+    align-items: center;
+    display: flex;
+    flex-direction: row;
+    font-size: $font_size-large;
+    padding: 30px 0;
+
+    &--selected > .year-navigation__year--lightgreen {
+      color: $color_brand-primary;
+      font-weight: bold;
+    }
+
+    &--selected > .year-navigation__year--darkgreen {
+      color: $color_font-dark;
+      font-weight: bold;
+    }
+
+    &--selected > .year-navigation__arrow {
+      @include media(medium-small) {
+        border-bottom: 20px solid $color_brand-primary;
+        border-left: 20px solid transparent;
+        border-right: 20px solid transparent;
+      }
+      border-bottom: 25px solid $color_brand-primary;
+      border-left: 30px solid transparent;
+      border-right: 30px solid transparent;
+      height: 0;
+      position: absolute;
+      top: 70px;
+      width: 0;
+    }
+  }
+}
+
+@include media(x-small) {
+  .year-navigation {
+    &__mobile {
+      background: $color_bg-medium;
+      border: 0;
+      color: $color_font-dark;
+      display: flex;
+      flex-direction: row;
+      font-size: $font_size-xlarge;
+      font-weight: bold;
+      margin: 0 auto;
+      outline: none;
+    }
+
+    &__desktop {
+      display: none;
+    }
+  }
+}
