+++ app/javascript/styles/components/_all.scss
@@ -10,3 +10,7 @@
 @import 'MapBox';
 @import 'visualizations/all';
 @import 'field/all';
+@import 'YearNav';
+@import 'Calendar';
+@import 'Login';
+@import 'Admin';
