+++ app/javascript/styles/partials/_vars.scss
@@ -21,6 +21,7 @@ $font_size-xsmall: .75rem;  // 12px
 $font_size-small: 1rem;  // 16px
 $font_size-medium: 1.25rem; // 20px
 $font_size-large: 1.875rem;  // 30px
+$font_size-xlarge: 2.5rem;
 
 $font_mod-small: .06125rem;
 $font_mod-medium: $font_mod-small * 2;
