+++ app/views/layouts/application.html.erb
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Datacommon</title>
+    <title>DataCommon</title>
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
